--- .vscode/settings.json
@@ -1,9 +1,0 @@
-{
-    "cSpell.words": [
-        "intro",
-        "Learn",
-        "pokecards",
-        "pokemon",
-        "pokemons"
-    ]
-}
