+++ frontendjr/src/App.css
@@ -10,15 +10,14 @@
   width: 8.75rem;
 }
 
-.barraIntro {
-  background-image: url("./assets/pokecards.jpg");
-  background-repeat: no-repeat;
-  background-size: 100%;
+.introBar {
   width: 100%;
+  background-image: url("./assets/pokecards.jpg");
+  background-size: 100%;
   height: 16.313rem;
 }
 
-.barraDeBusca {
+.searchBar {
   background-image: url("./assets/lupa.svg");
   background-repeat: no-repeat;
   background-position: 97% 50%;
@@ -36,26 +35,23 @@
   outline: 0;
 }
 
-.pagina {
+.page {
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  padding: 10px;
+  padding: 0.625rem;
 }
 
 .cards {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
-  grid-gap: 37px;
-  margin-top: 86px;
-  margin-bottom: 116px;
-  margin-left: 168px;
+  grid-gap: 2.313rem;
+  margin-top: 5.375rem;
+  margin-bottom: 7.25rem;
+  margin-left: 10.5rem;
 }
 
 .pokemonCards {
-  width: 232px;
-  height: 265px;
+  width: 14.5rem;
+  height: 16.563rem;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -65,16 +61,16 @@
 }
 
 .pokemonImg {
-  margin-top: 30px;
-  width: 64px;
-  height: 63px;
+  margin-top: 1.875rem;
+  width: 4.375rem;
+  height: 4.313rem;
 }
 
 
 .infoCards {
   display: flex;
-  padding: 10px;
-  grid-gap: 10px;
+  padding: 0.625rem;
+  grid-gap: 0.625rem;
   align-items: center;
   justify-content: center;
 }
@@ -96,7 +92,7 @@
   align-items: center;
   justify-content: space-evenly;
   width: 50%;
-  height: 43px;
+  height: 2.688rem;
   background-color: #2726260f;
   border: none;
   cursor: pointer;
