+++ frontendjr/src/api.js
@@ -1,4 +1,4 @@
-export const buscarPokemon = async (pokemon) => {
+export const searchPokemon = async (pokemon) => {
     try {
         let url = `https://pokeapi.co/api/v2/pokemon/${pokemon}`
         const response = await fetch(url);
