+++ frontendjr/src/components/Pokedex.js
@@ -7,18 +7,15 @@ const Pokedex = (props) => {
     const { pokemons } = props;
 
     return (
-        <div>
-            <div className="pagina">
+            <div className="page">
                 <div className="cards">
                     {pokemons.map((pokemon, idx) => {
                         return (
                             <Pokemon pokemon={pokemon} key={pokemon.name} />
-                            // <div key={pokemon.name}>#{idx + 1}: {pokemon.name}</div>
                         )
                     })}
                 </div>
             </div>
-        </div>
     )
 }
